removed old unit test + fix import error
